+++ testproject/django_file_form_example/models.py
@@ -9,12 +9,21 @@ class Example(models.Model):
     title = models.CharField(max_length=255)
     input_file = models.FileField(max_length=255, upload_to="example", storage=fs)
 
+    def __str__(self):
+        return self.title
+
 
 class Example2(models.Model):
     title = models.CharField(max_length=255)
 
+    def __str__(self):
+        return self.title
+
 
 class ExampleFile(models.Model):
+    class Meta:
+        ordering = ["input_file"]
+
     fs = FileSystemStorage(location=settings.MEDIA_ROOT)
 
     example = models.ForeignKey(