+++ testproject/django_file_form_example/tests/live/test_admin.py
@@ -3,7 +3,7 @@ from django.core.files.base import ContentFile
 from django_file_form_example.models import Example
 from django_file_form_example.tests.utils.admin_page import AdminPage
 from django_file_form_example.tests.utils.base_live_testcase import BaseLiveTestCase
-from django_file_form_example.tests.utils.test_utils import read_file
+from django_file_form_example.tests.utils.test_utils import read_file, get_random_id
 
 
 class AdminTestCase(BaseLiveTestCase):
@@ -29,7 +29,7 @@ class AdminTestCase(BaseLiveTestCase):
     def test_edit_record(self):
         example = Example.objects.create(
             title="title1",
-            input_file=ContentFile("original", "test1.txt"),
+            input_file=ContentFile("original", get_random_id()),
         )
 
         try:
@@ -51,4 +51,3 @@ class AdminTestCase(BaseLiveTestCase):
             self.assertEqual(read_file(example.input_file), b"new_content")
         finally:
             example.input_file.delete()
-